refactor: 하이라이팅 코드 리팩토링
